Move theme buttons to ThemeGrid component
